+++ changelog.md
@@ -1,6 +1,8 @@
-### 0.5.2, 2023 Aug 6
+### 0.6.0, 2023 Aug 6
 
 - When multiple object files have the same filename, disambiguate them (output their folder name in that case too).
+- Object file sizes are counted better, in the case where some of their contributions are without symbols.
+- Sizer output now also lists object files by data size.
 - Slightly improve performance.
 
 ### 0.5.1, 2023 Aug 5